Title Read a municipal data file to a tibble

This function is a wrapper around readxl::readexcel(), reading a specific municipal data file for a specific year and a specific data domain. Its added value is in its use of row_to_names_fill() and its pre-defined parameters for every year and its specific quirks in the Excel headers.

Usage

read_cbs_muni(path, year, data_domain = c("physical", "budget"))

Arguments

path

A character vector of length 1, denoting the local file path to the municipal data file. A full list of available files by the CBS is at https://www.cbs.gov.il/he/publications/Pages/2019/%D7%94%D7%A8%D7%A9%D7%95%D7%99%D7%95%D7%AA-%D7%94%D7%9E%D7%A7%D7%95%D7%9E%D7%99%D7%95%D7%AA-%D7%91%D7%99%D7%A9%D7%A8%D7%90%D7%9C-%D7%A7%D7%95%D7%91%D7%A6%D7%99-%D7%A0%D7%AA%D7%95%D7%A0%D7%99%D7%9D-%D7%9C%D7%A2%D7%99%D7%91%D7%95%D7%93-1999-2017.aspx

year

A numeric vector of length 1 denoting which year the data file pointed in path is for. Currently supporting only 2016 and later, since before 2016 regional councils are split from cities and local councils.

data_domain

A character vector of length 1, one of c("physical", "budget"). Every Excel municipal data file has a few different data domains, most notably physical and population data, and budget data.

Value

A tibble with municipal data for a specific year, where every row is a municipality and every column is a different variable for this municipality in that year. Be advised all columns are of type character, so you nee to parse the data types yourself at will.

Examples

df <- read_cbs_muni(
  system.file("extdata", "p_libud_2021.xlsx", package = "il.cbs.muni"),
  year = 2021,
  data_domain = "physical"
)

df |>
  head() |>
  str()
#> tibble [6 × 300] (S3: tbl_df/tbl/data.frame)
#>  $ שם  הרשות                                                                                                                                          : chr [1:6] "שנת עדכון: 2021\r\n(אלא אם כן צוין אחרת)" "אום אל-פחם" "אופקים" "אור יהודה" ...
#>  $ כללי_סמל הרשות                                                                                                                                     : chr [1:6] NA "2710" "0031" "2400" ...
#>  $ כללי_מחוז                                                                                                                                          : chr [1:6] NA "חיפה" "הדרום" "תל אביב" ...
#>  $ כללי_מעמד מוניציפלי                                                                                                                                : chr [1:6] NA "עירייה" "עירייה" "עירייה" ...
#>  $ כללי_מרחק מגבול מחוז תל אביב (ק"מ)                                                                                                                 : chr [1:6] "2020" "60.581242748414603" "87.130838665736604" "0" ...
#>  $ כללי_שנת קבלת מעמד מוניציפלי                                                                                                                       : chr [1:6] NA "1984" "1995" "1988" ...
#>  $ כללי_מרחב ימי, רשויות מקומיות הגובלות בחוף הים                                                                                                     : chr [1:6] NA "0" "0" "0" ...
#>  $ כללי_מספר חברי מועצה                                                                                                                               : chr [1:6] NA "17" "15" "15" ...
#>  $ כללי_סמל ועדת תכנון ובנייה                                                                                                                         : chr [1:6] NA "354" "601" "510" ...
#>  $ כללי_שם ועדת תכנון ובנייה                                                                                                                          : chr [1:6] NA "עירון" "אופקים" "אור יהודה" ...
#>  $ כללי_שטח 
#> (קמ"ר)                                                                                                                                : chr [1:6] NA "26.028286285781743" "16.352399436075068" "6.7300239620150855" ...
#>  $ דמוגרפיה_צפיפות אוכלוסייה לקמ''ר ביישובים שמנו 5,000 תושבים ויותר                                                                                  : chr [1:6] NA "2221.1" "2079.1" "5470.2" ...
#>  $ דמוגרפיה_סה"כ  אוכלוסייה בסוף השנה                                                                                                                 : chr [1:6] NA "57677" "33999" "36815" ...
#>  $ דמוגרפיה_יהודים ואחרים (אחוזים)                                                                                                                    : chr [1:6] NA "-" "99.3" "99.8" ...
#>  $ דמוגרפיה_יהודים 
#> (אחוזים מתוך יהודים ואחרים)                                                                                                    : chr [1:6] NA "-" "93.4" "97.6" ...
#>  $ דמוגרפיה_ערבים (אחוזים)                                                                                                                            : chr [1:6] NA "99.9" "-" "-" ...
#>  $ דמוגרפיה_מוסלמים (אחוזים מתוך האוכלוסייה הערבית)                                                                                                   : chr [1:6] NA "100" "-" "-" ...
#>  $ דמוגרפיה_נוצרים (אחוזים מתוך האוכלוסייה הערבית)                                                                                                    : chr [1:6] NA "-" "-" "-" ...
#>  $ דמוגרפיה_דרוזים (אחוזים מתוך האוכלוסייה הערבית)                                                                                                    : chr [1:6] NA "-" "-" "-" ...
#>  $ דמוגרפיה_סה"כ גברים בסוף השנה                                                                                                                      : chr [1:6] NA "29408" "16805" "18284" ...
#>  $ דמוגרפיה_סה"כ נשים בסוף השנה                                                                                                                       : chr [1:6] NA "28269" "17194" "18531" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 4-0                                                                                                         : chr [1:6] NA "9.6832359519392472" "13.926880202358893" "8.5590112725791112" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 9-5                                                                                                         : chr [1:6] NA "9.457842814293393" "11.247389629106738" "8.2765177237539049" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 14-10                                                                                                       : chr [1:6] NA "10.114950500199386" "8.4414247477866997" "7.3231019964688304" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 19-15                                                                                                       : chr [1:6] NA "10.602146436187736" "6.7766699020559429" "6.8993616732310201" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 29-20                                                                                                       : chr [1:6] NA "19.054389097907311" "15.579869996176358" "13.543392638870026" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 44-30                                                                                                       : chr [1:6] NA "17.913553062745983" "19.2858613488632" "20.119516501426048" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 59-45                                                                                                       : chr [1:6] NA "15.463703035872184" "12.262125356628136" "16.24066277332609" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 64-60                                                                                                       : chr [1:6] NA "2.862492848102363" "3.8324656607547283" "5.2505772103762052" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 65 ומעלה                                                                                                    : chr [1:6] NA "4.8494200461189036" "8.647313156269302" "13.787858209968762" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 17-0                                                                                                        : chr [1:6] NA "35.38325502366628" "37.651107385511338" "28.279234007877225" ...
#>  $ דמוגרפיה_אחוז באוכלוסייה בסוף השנה_בני 75 ומעלה                                                                                                    : chr [1:6] NA "1.6488374915477573" "3.2736256948733788" "5.049572185250577" ...
#>  $ דמוגרפיה_יחס תלות (ל-1,000 תושבים בלתי תלויים)                                                                                                     : chr [1:6] NA "809" "962" "813" ...
#>  $ דמוגרפיה_לידות חי                                                                                                                                  : chr [1:6] NA "1130" "992" "639" ...
#>  $ דמוגרפיה_שיעור פריון כולל                                                                                                                          : chr [1:6] NA "2.54" "3.95" "2.58" ...
#>  $ דמוגרפיה_פטירות תינוקות (ממוצע)                                                                                                                    : chr [1:6] "2021-2017" "4.4000000000000004" "3.8" "0.6" ...
#>  $ דמוגרפיה_שיעור פטירות תינוקות ל-1,000 לידות חי                                                                                                     : chr [1:6] "2021-2017" "3.9" "( 4.2)" ".." ...
#>  $ דמוגרפיה_פטירות                                                                                                                                    : chr [1:6] NA "194" "208" "233" ...
#>  $ דמוגרפיה_שיעור פטירות מתוקנן ל-1,000 תושבים                                                                                                        : chr [1:6] "2021-2017" "6.14" "6.56" "5.18" ...
#>  $ דמוגרפיה_ריבוי טבעי
#> סה"כ                                                                                                                        : chr [1:6] NA "936" "784" "406" ...
#>  $ דמוגרפיה_ריבוי טבעי  ל-1,000 תושבים                                                                                                                : chr [1:6] NA "16.355960263169514" "23.55998717634187" "11.034928999156906" ...
#>  $ דמוגרפיה_אחוז גידול האוכלוסייה לעומת השנה הקודמת                                                                                                   : chr [1:6] NA "1.5856352146045083" "4.4364407952287905" "0.12222272320148875" ...
#>  $ דמוגרפיה_אחוז עולי 1990+ מסך האוכלוסייה                                                                                                            : chr [1:6] NA "3.1208280597118436E-2" "16.885790758551721" "13.91280727964145" ...
#>  $ דמוגרפיה_השתקעות ראשונה של עולים                                                                                                                   : chr [1:6] NA ".." "42" "29" ...
#>  $ דמוגרפיה_אחוז עולים בהשתקעות ראשונה_בני 17-0                                                                                                       : chr [1:6] NA ".." "28.571428571428569" "37.931034482758619" ...
#>  $ דמוגרפיה_אחוז עולים בהשתקעות ראשונה_בני 44-18                                                                                                      : chr [1:6] NA ".." "47.619047619047613" "41.379310344827587" ...
#>  $ דמוגרפיה_אחוז עולים בהשתקעות ראשונה_בני 64-45                                                                                                      : chr [1:6] NA ".." "19.047619047619047" "17.241379310344829" ...
#>  $ דמוגרפיה_אחוז עולים בהשתקעות ראשונה_בני 65 ומעלה                                                                                                   : chr [1:6] NA ".." "4.7619047619047619" "3.4482758620689653" ...
#>  $ דמוגרפיה_מאזן הגירה ביישוב
#> סה"כ                                                                                                                 : chr [1:6] NA "-70" "692" "-390" ...
#>  $ דמוגרפיה_מאזן הגירה פנימית                                                                                                                         : chr [1:6] NA "-68" "637" "-429" ...
#>  $ דמוגרפיה_נכנסים מיישובים אחרים
#> _סה"כ                                                                                                            : chr [1:6] NA "266" "1741" "929" ...
#>  $ דמוגרפיה_נכנסים מיישובים אחרים
#> _גברים                                                                                                           : chr [1:6] NA "81" "879" "464" ...
#>  $ דמוגרפיה_נכנסים מיישובים אחרים
#> _נשים                                                                                                            : chr [1:6] NA "185" "862" "465" ...
#>  $ דמוגרפיה_נכנסים מיישובים אחרים
#> _עד גיל 14                                                                                                       : chr [1:6] NA "75" "568" "261" ...
#>  $ דמוגרפיה_נכנסים מיישובים אחרים
#> _גיל 29-15                                                                                                       : chr [1:6] NA "127" "524" "209" ...
#>  $ דמוגרפיה_נכנסים מיישובים אחרים
#> _גיל 64-30                                                                                                       : chr [1:6] NA "63" "609" "395" ...
#>  $ דמוגרפיה_נכנסים מיישובים אחרים
#> _גיל 65 ומעלה                                                                                                    : chr [1:6] NA "1" "40" "64" ...
#>  $ דמוגרפיה_יוצאים ליישובים אחרים
#> _סה"כ                                                                                                            : chr [1:6] NA "334" "1104" "1358" ...
#>  $ דמוגרפיה_יוצאים ליישובים אחרים
#> _גברים                                                                                                           : chr [1:6] NA "128" "577" "682" ...
#>  $ דמוגרפיה_יוצאים ליישובים אחרים
#> _נשים                                                                                                            : chr [1:6] NA "206" "527" "676" ...
#>  $ דמוגרפיה_יוצאים ליישובים אחרים
#> _עד גיל 14                                                                                                       : chr [1:6] NA "80" "346" "389" ...
#>  $ דמוגרפיה_יוצאים ליישובים אחרים
#> _גיל 29-15                                                                                                       : chr [1:6] NA "163" "349" "355" ...
#>  $ דמוגרפיה_יוצאים ליישובים אחרים
#> _גיל 64-30                                                                                                       : chr [1:6] NA "87" "364" "523" ...
#>  $ דמוגרפיה_יוצאים ליישובים אחרים
#> _גיל 65 ומעלה                                                                                                    : chr [1:6] NA "4" "45" "91" ...
#>  $ דמוגרפיה_הגירה בתוך הרשות המקומית                                                                                                                  : chr [1:6] NA "951" "2284" "1280" ...
#>  $ דמוגרפיה_נישאים_סה"כ גברים ונשים                                                                                                                   : chr [1:6] "2020" "710" "398" "266" ...
#>  $ דמוגרפיה_נישאים_שיעור ל-1,000 תושבים לא נשואים בני 15 ומעלה                                                                                        : chr [1:6] "2020" "45" "41.3" "21" ...
#>  $ דמוגרפיה_מתגרשים_סה"כ גברים ונשים                                                                                                                  : chr [1:6] "2020" "90" "158" "147" ...
#>  $ דמוגרפיה_מתגרשים_שיעור ל-1,000 תושבים נשואים בני 15 ומעלה                                                                                          : chr [1:6] "2020" "3.8" "13.5" "9.6" ...
#>  $ דמוגרפיה_אוכלוסיית דיור משותף (אחוז מתוך אוכלוסיית הרשות)_סך מוסדות דיור משותף ברשות                                                               : chr [1:6] "2020" ".." "6" "8" ...
#>  $ דמוגרפיה_אוכלוסיית דיור משותף (אחוז מתוך אוכלוסיית הרשות)_סך דיירי דיור משותף ברשות                                                                : chr [1:6] "2020" ".." "559" "488" ...
#>  $ דמוגרפיה_אוכלוסיית דיור משותף (אחוז מתוך אוכלוסיית הרשות)_אחוז דיירים בבתי אבות ובדיור מוגן                                                        : chr [1:6] "2020" ".." "0.87" "0.46" ...
#>  $ דמוגרפיה_אוכלוסיית דיור משותף (אחוז מתוך אוכלוסיית הרשות)_אחוז דיירים במוסדות חינוך                                                                : chr [1:6] "2020" ".." "0.57999999999999996" "-" ...
#>  $ דמוגרפיה_אוכלוסיית דיור משותף (אחוז מתוך אוכלוסיית הרשות)_אחוז דיירים במוסדות דת                                                                   : chr [1:6] "2020" ".." "0.17" "0.46" ...
#>  $ דמוגרפיה_אוכלוסיית דיור משותף (אחוז מתוך אוכלוסיית הרשות)_אחוז דיירים במוסדות אשפוז כולל הוסטלים                                                   : chr [1:6] "2020" ".." "-" "0.41" ...
#>  $ דמוגרפיה_אוכלוסיית דיור משותף (אחוז מתוך אוכלוסיית הרשות)_אחוז דיירים במוסדות אחרים                                                                : chr [1:6] "2020" ".." "0.09" "-" ...
#>  $ בריאות_תוחלת חיים
#> בלידה סך הכל_בערים המונות 50,000 תושבים ויותר                                                                                 : chr [1:6] "2017 - 2021" "78.774353753574403" ".." ".." ...
#>  $ בריאות_מספר מקרי סוכרת ממוצע לשנה_בעיריות ומועצות מקומיות המונות 10,000 תושבים ויותר                                                               : chr [1:6] "2019-2017" "3822" "1828.6666666666667" "2838.6666666666665" ...
#>  $ בריאות_שיעור מקרי סוכרת מתוקנן ל-1,000 תושבים
#> _בעיריות ומועצות מקומיות המונות 10,000 תושבים ויותר                                               : chr [1:6] "2019-2017" "109.29215413642061" "68.676340518052456" "64.861141396707694" ...
#>  $ בריאות_עודף משקל והשמנה, כיתה א 
#> _אחוזים
#> בעיריות ומועצות מקומיות המונות 10,000 תושבים ויותר                                                  : chr [1:6] "תשפ\"א 2020/21" "26" "21" "29" ...
#>  $ בריאות_עודף משקל והשמנה, כיתה ז_אחוזים
#> בעיריות ומועצות מקומיות המונות 10,000 תושבים ויותר                                                       : chr [1:6] NA "46" "38" "42" ...
#>  $ בריאות_מספר מקרי סרטן ממוצע לשנה, גברים_בעיריות ומועצות מקומיות המונות 20,000 תושבים ויותר                                                         : chr [1:6] "2018-2014" "40.200000000000003" "38.799999999999997" "57.6" ...
#>  $ בריאות_שיעור סרטן מכל הסוגים מתוקנן ל-100,000 תושבים, גברים_בעיריות ומועצות מקומיות המונות 20,000 תושבים ויותר                                     : chr [1:6] "2018-2014" "312.98659639567387" "343.084734280276" "305.99580704036714" ...
#>  $ בריאות_מספר מקרי סרטן ממוצע לשנה, נשים_בעיריות ומועצות מקומיות המונות 20,000 תושבים ויותר                                                          : chr [1:6] "2018-2014" "37.6" "47.6" "60.2" ...
#>  $ בריאות_שיעור סרטן מכל הסוגים מתוקנן ל-100,000 תושבים, נשים_בעיריות ומועצות מקומיות המונות 20,000 תושבים ויותר                                      : chr [1:6] "2018-2014" "219.07898197907264" "353.66793819192651" "275.13977473170206" ...
#>  $ שכר ורווחה_מספר מקבלי דמי אבטלה 
#> (סה"כ שנתי)                                                                                                    : chr [1:6] NA "5143" "3134" "4147" ...
#>  $ שכר ורווחה_אחוז מקבלי דמי אבטלה מבני 67-20
#> (שנתי)                                                                                               : chr [1:6] NA "15.74419886120125" "17.421757740841628" "19.402077290165622" ...
#>  $ שכר ורווחה_מקבלי דמי אבטלה 
#> (ממוצע חודשי)_סה"כ                                                                                                  : chr [1:6] NA "2325" "1236" "1615" ...
#>  $ שכר ורווחה_מקבלי דמי אבטלה 
#> (ממוצע חודשי)_גברים                                                                                                 : chr [1:6] NA "1459" "490" "739" ...
#>  $ שכר ורווחה_אחוז מקבלי דמי אבטלה מבני 67-20
#> (ממוצע חודשי)                                                                                        : chr [1:6] NA "7.2252089872276954" "7.005214237134437" "7.5460237360994302" ...
#>  $ שכר ורווחה_גיל ממוצע של מקבלי דמי אבטלה
#> (לא כולל חיילים)                                                                                        : chr [1:6] NA "37" "38" "41" ...
#>  $ שכר ורווחה_דמי אבטלה ממוצעים ליום
#> (ש"ח)                                                                                                         : chr [1:6] NA "151" "164" "178" ...
#>  $ שכר ורווחה_ימי זכאות לדמי אבטלה
#> (ממוצע שנתי)                                                                                                    : chr [1:6] NA "123" "121" "129" ...
#>  $ שכר ורווחה_ניצול ימי אבטלה בשנה 
#> (ממוצע שנתי)                                                                                                   : chr [1:6] NA "131" "110" "109" ...
#>  $ שכר ורווחה_ניצול ימי אבטלה בשנה 
#> (חציון שנתי)                                                                                                   : chr [1:6] NA "141" "113" "104" ...
#>  $ שכר ורווחה_מקבלי הכשרה מקצועית
#> סה"כ                                                                                                             : chr [1:6] NA ".." ".." ".." ...
#>  $ שכר ורווחה_מקבלי קצבאות זקנה ושאירים
#> סה"כ
#> (סוף שנה)                                                                                          : chr [1:6] NA "2613" "3194" "5074" ...
#>  $ שכר ורווחה_אחוז מקבלי השלמת הכנסה ממקבלי קצבאות זקנה ושאירים
#> (סוף שנה)                                                                          : chr [1:6] NA "24.531190202831993" "36.380713838447086" "13.657863618446985" ...
#>  $ שכר ורווחה_מקבלי גמלת הבטחת הכנסה
#> (נפשות, סוף שנה)                                                                                              : chr [1:6] NA "4292" "1383" "753" ...
#>   [list output truncated]